All Client/Dog Functionality Transferred from CoreData to CloudKit
Creating, Retrieving, Updating, and Deleting of Client and Dog records works. If Client is deleted so is all of their owned dogs. Pulling from Cloudkit is kept to a minimum and only done upon first launch, after that Cloudkit is only used for saving/deleting while the UI is updated via array of Clients and array of Dogs.